monitor/requirements.txt
pylightning>=0.0.6

